Stress Element Diagram Rotation
Introduced a point rotation function to rotate stress element points for the diagram. Axis reference line how rotates and translates to look more natural with the rotation of the element. Second stress arrow was added and a label was added to the stress element for shear stress. Stress element is how callable and is able to take an angle of rotation in degrees.